+++ src/zenml/integrations/kserve/custom_deployer/zenml_custom_model.py
@@ -73,7 +73,7 @@ class ZenMLCustomModel(kserve.Model):  # type: ignore[misc]
 
         """
         try:
-            from zenml.utils.materializer_utils import load_model_from_metadata
+            from zenml.utils.artifact_utils import load_model_from_metadata
 
             self.model = load_model_from_metadata(self.model_uri)
         except Exception as e:
