+++ src/zenml/integrations/kserve/steps/kserve_deployer.py
@@ -44,7 +44,7 @@ from zenml.steps import (
 )
 from zenml.steps.step_context import StepContext
 from zenml.utils import io_utils, source_utils
-from zenml.utils.materializer_utils import save_model_metadata
+from zenml.utils.artifact_utils import save_model_metadata
 
 logger = get_logger(__name__)
