+++ src/zenml/integrations/seldon/custom_deployer/zenml_custom_model.py
@@ -74,7 +74,7 @@ class ZenMLCustomModel:
 
         """
         try:
-            from zenml.utils.materializer_utils import load_model_from_metadata
+            from zenml.utils.artifact_utils import load_model_from_metadata
 
             self.model = load_model_from_metadata(self.model_uri)
         except Exception as e:
