+++ src/zenml/integrations/seldon/steps/seldon_deployer.py
@@ -49,7 +49,7 @@ from zenml.steps import (
 )
 from zenml.steps.step_context import StepContext
 from zenml.utils import io_utils, source_utils
-from zenml.utils.materializer_utils import save_model_metadata
+from zenml.utils.artifact_utils import save_model_metadata
 
 logger = get_logger(__name__)
